refactor(lowlevel/envelope): move model definitions out
